Fixing zero onhand

+++ src/main/java/com/olteanuflorin86/msscbeerinventoryservicev1/services/AllocationServiceImpl.java
@@ -78,7 +78,9 @@ public class AllocationServiceImpl implements AllocationService {
                 beerOrderLine.setQuantityAllocated(allocatedQty + inventory);
                 beerInventory.setQuantityOnHand(0);
 
-                beerInventoryRepository.delete(beerInventory);
+                if (beerInventory.getQuantityOnHand() == 0) {
+                    beerInventoryRepository.delete(beerInventory);
+                }
             }
         });
